Source a .vimrc.local
I've added sourcing a .vimrc.local so that those things that need to be used
only in certain environments don't get in the way of the overall files, such as
setting up the right author and or email to be used, definitely going to be
different at work and at home.

Also set up some defaults for snipmate.