Fix an issue with clang_complete
When clang_complete encounters an error it is generally nice to have it
show up in the quick fix window. However when that happens it will no
longer let me go into insert mode and thereby doesn't allow me to do
anything to fix the issues. This requires a full restart of vim which I
do not like. For now we just disable the quickfix window.